+++ src/Point2D.java
@@ -1,20 +1,20 @@
 public class Point2D {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point2D(int x , int y){
+    public Point2D(double x , double y){
         this.x=x;
         this.y=y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setLocation(int x, int y){
+    public void setLocation(double x, double y){
         this.x=x;
         this.y=y;
     }
